# src/main/java/com/zjsu/syt/course/model/ScheduleSlot.java
package com.zjsu.syt.course.model;

import java.time.LocalTime;

public class ScheduleSlot {
    private String dayOfWeek;
    private LocalTime startTime;
    private LocalTime endTime;
    private Integer expectedAttendance;

    public ScheduleSlot() {}

    public ScheduleSlot(String dayOfWeek, LocalTime startTime, LocalTime endTime, Integer expectedAttendance) {
        this.dayOfWeek = dayOfWeek;
        this.startTime = startTime;
        this.endTime = endTime;
        this.expectedAttendance = expectedAttendance;
    }

    // Getters and Setters
    public String getDayOfWeek() { return dayOfWeek; }
    public void setDayOfWeek(String dayOfWeek) { this.dayOfWeek = dayOfWeek; }

    public LocalTime getStartTime() { return startTime; }
    public void setStartTime(LocalTime startTime) { this.startTime = startTime; }

    public LocalTime getEndTime() { return endTime; }
    public void setEndTime(LocalTime endTime) { this.endTime = endTime; }

    public Integer getExpectedAttendance() { return expectedAttendance; }
    public void setExpectedAttendance(Integer expectedAttendance) { this.expectedAttendance = expectedAttendance; }
}